+++ subgraph/schema.graphql
@@ -14,6 +14,7 @@ type Foundation @entity {
   vault: Vault!
   deposits: [Deposit!]! @derivedFrom(field: "foundation")
   amountDeposited: BigInt!
+  shares: BigInt!
   lockedUntil: BigInt!
   createdAt: BigInt!
 }
@@ -31,7 +32,7 @@ type Deposit @entity {
 }
 
 type Claimer @entity {
-  id: Bytes!
+  id: ID!
   owner: Bytes!
   vault: Vault!
   deposits: [Deposit!]! @derivedFrom(field: "claimer")
