+++ subgraph/tests/integration.test.ts
@@ -1,11 +1,20 @@
-import { ethereum, Address, Bytes, BigInt } from '@graphprotocol/graph-ts';
-import { log } from 'matchstick-as/assembly/log';
+import { Address, Bytes, BigInt } from '@graphprotocol/graph-ts';
 import {
   test,
   assert,
   newMockEvent,
   clearStore,
 } from 'matchstick-as/assembly/index';
+
+import {
+  newBytes,
+  newI32,
+  newBool,
+  newAddress,
+  newString,
+  donationId,
+  createDeposit,
+} from './helpers';
 
 import {
   handleDepositMinted,
@@ -52,373 +61,7 @@ const MOCK_ADDRESS_1 = '0xC80B3caAd6d2DE80Ac76a41d5F0072E36D2519Cd'.toLowerCase(
 const MOCK_ADDRESS_2 = '0xE80B3caAd6d2DE80Ac76a41d5F0072E36D2519Ce'.toLowerCase();
 const TREASURY_ADDRESS = '0x4940c6e628da11ac0bdcf7f82be8579b4696fa33';
 
-test('handleInitDeposit creates a DepositOperation', () => {
-  clearStore();
-
-  let mockEvent = newMockEvent();
-  const event = new InitDepositStable(
-    mockEvent.address,
-    mockEvent.logIndex,
-    mockEvent.transactionLogIndex,
-    mockEvent.logType,
-    mockEvent.block,
-    mockEvent.transaction,
-    mockEvent.parameters,
-  );
-  event.parameters = new Array();
-
-  const operator = newAddress('operator', MOCK_ADDRESS_1);
-  const idx = newI32('idx', 1);
-  const underlyingAmount = newI32('underlyingAmount', 1);
-  const ustAmount = newI32('ustAmount', 1);
-
-  event.parameters.push(operator);
-  event.parameters.push(idx);
-  event.parameters.push(ustAmount);
-  event.parameters.push(underlyingAmount);
-
-  handleInitDeposit(event);
-
-  assert.fieldEquals('DepositOperation', MOCK_ADDRESS_1, 'idx', '1');
-  assert.fieldEquals(
-    'DepositOperation',
-    MOCK_ADDRESS_1,
-    'underlyingAmount',
-    '1',
-  );
-  assert.fieldEquals('DepositOperation', MOCK_ADDRESS_1, 'ustAmount', '1');
-});
-
-test('RearrangeDepositOperation updates the DepositOperation idx', () => {
-  clearStore();
-
-  let mockDepositEvent = newMockEvent();
-  const depositEvent = new InitDepositStable(
-    mockDepositEvent.address,
-    mockDepositEvent.logIndex,
-    mockDepositEvent.transactionLogIndex,
-    mockDepositEvent.logType,
-    mockDepositEvent.block,
-    mockDepositEvent.transaction,
-    mockDepositEvent.parameters,
-  );
-  depositEvent.parameters = new Array();
-
-  let operator = newAddress('operator', MOCK_ADDRESS_1);
-  let idx = newI32('idx', 1);
-  let underlyingAmount = newI32('underlyingAmount', 1);
-  let ustAmount = newI32('ustAmount', 1);
-
-  depositEvent.parameters.push(operator);
-  depositEvent.parameters.push(idx);
-  depositEvent.parameters.push(ustAmount);
-  depositEvent.parameters.push(underlyingAmount);
-
-  handleInitDeposit(depositEvent);
-
-  let mockDepositEvent1 = newMockEvent();
-  const depositEvent1 = new InitDepositStable(
-    mockDepositEvent1.address,
-    mockDepositEvent1.logIndex,
-    mockDepositEvent1.transactionLogIndex,
-    mockDepositEvent1.logType,
-    mockDepositEvent1.block,
-    mockDepositEvent1.transaction,
-    mockDepositEvent1.parameters,
-  );
-  depositEvent1.parameters = new Array();
-
-  operator = newAddress('operator', MOCK_ADDRESS_2);
-  idx = newI32('idx', 2);
-
-  depositEvent1.parameters.push(operator);
-  depositEvent1.parameters.push(idx);
-  depositEvent1.parameters.push(ustAmount);
-  depositEvent1.parameters.push(underlyingAmount);
-
-  handleInitDeposit(depositEvent1);
-
-  let mockEvent = newMockEvent();
-  const event = new RearrangeDepositOperation(
-    mockEvent.address,
-    mockEvent.logIndex,
-    mockEvent.transactionLogIndex,
-    mockEvent.logType,
-    mockEvent.block,
-    mockEvent.transaction,
-    mockEvent.parameters,
-  );
-  event.parameters = new Array();
-
-  const operatorFrom = newAddress('operatorFrom', MOCK_ADDRESS_2);
-  const operatorTo = newAddress('operatorTo', MOCK_ADDRESS_1);
-  idx = newI32('idx', 1);
-
-  event.parameters.push(operatorFrom);
-  event.parameters.push(operatorTo);
-  event.parameters.push(idx);
-
-  assert.fieldEquals('DepositOperation', MOCK_ADDRESS_2, 'idx', '2');
-  assert.fieldEquals('DepositOperation', MOCK_ADDRESS_1, 'idx', '1');
-
-  handleRearrangeDeposit(event);
-
-  assert.fieldEquals('DepositOperation', MOCK_ADDRESS_2, 'idx', '1');
-  assert.fieldEquals('DepositOperation', MOCK_ADDRESS_1, 'idx', '1');
-});
-
-test('handleFinishDeposit deletes the DepositOperation', () => {
-  clearStore();
-
-  const depositOperation = new DepositOperation(MOCK_ADDRESS_1);
-  depositOperation.save();
-
-  let mockEvent = newMockEvent();
-  const event = new FinishDepositStable(
-    mockEvent.address,
-    mockEvent.logIndex,
-    mockEvent.transactionLogIndex,
-    mockEvent.logType,
-    mockEvent.block,
-    mockEvent.transaction,
-    mockEvent.parameters,
-  );
-  event.parameters = new Array();
-
-  const operator = newAddress('operator', MOCK_ADDRESS_1);
-
-  event.parameters.push(operator);
-
-  handleFinishDeposit(event);
-
-  assert.notInStore('DepositOperation', MOCK_ADDRESS_1);
-});
-
-test('handleInitRedeem creates a RedeemOperation', () => {
-  clearStore();
-
-  let mockEvent = newMockEvent();
-  const event = new InitRedeemStable(
-    mockEvent.address,
-    mockEvent.logIndex,
-    mockEvent.transactionLogIndex,
-    mockEvent.logType,
-    mockEvent.block,
-    mockEvent.transaction,
-    mockEvent.parameters,
-  );
-  event.parameters = new Array();
-
-  const operator = newAddress('operator', MOCK_ADDRESS_1);
-  const idx = newI32('idx', 1);
-  const aUstAmount = newI32('aUstAmount', 1);
-
-  event.parameters.push(operator);
-  event.parameters.push(idx);
-  event.parameters.push(aUstAmount);
-
-  handleInitRedeem(event);
-
-  assert.fieldEquals('RedeemOperation', MOCK_ADDRESS_1, 'aUstAmount', '1');
-});
-
-test('RearrangeRedeemOperation updates the RedeemOperation idx', () => {
-  clearStore();
-
-  let mockRedeemEvent = newMockEvent();
-  const redeemEvent = new InitRedeemStable(
-    mockRedeemEvent.address,
-    mockRedeemEvent.logIndex,
-    mockRedeemEvent.transactionLogIndex,
-    mockRedeemEvent.logType,
-    mockRedeemEvent.block,
-    mockRedeemEvent.transaction,
-    mockRedeemEvent.parameters,
-  );
-  redeemEvent.parameters = new Array();
-
-  let operator = newAddress('operator', MOCK_ADDRESS_1);
-  let aUstAmount = newI32('aUstAmount', 1);
-  let idx = newI32('idx', 1);
-
-  redeemEvent.parameters.push(operator);
-  redeemEvent.parameters.push(idx);
-  redeemEvent.parameters.push(aUstAmount);
-
-  handleInitRedeem(redeemEvent);
-
-  let mockRedeemEvent1 = newMockEvent();
-  const redeemEvent1 = new InitRedeemStable(
-    mockRedeemEvent1.address,
-    mockRedeemEvent1.logIndex,
-    mockRedeemEvent1.transactionLogIndex,
-    mockRedeemEvent1.logType,
-    mockRedeemEvent1.block,
-    mockRedeemEvent1.transaction,
-    mockRedeemEvent1.parameters,
-  );
-  redeemEvent1.parameters = new Array();
-
-  operator = newAddress('operator', MOCK_ADDRESS_2);
-  idx = newI32('idx', 2);
-
-  redeemEvent1.parameters.push(operator);
-  redeemEvent1.parameters.push(idx);
-  redeemEvent1.parameters.push(aUstAmount);
-
-  handleInitRedeem(redeemEvent1);
-
-  const mockEvent = newMockEvent();
-  const event = new RearrangeRedeemOperation(
-    mockEvent.address,
-    mockEvent.logIndex,
-    mockEvent.transactionLogIndex,
-    mockEvent.logType,
-    mockEvent.block,
-    mockEvent.transaction,
-    mockEvent.parameters,
-  );
-  event.parameters = new Array();
-
-  const operatorFrom = newAddress('operatorFrom', MOCK_ADDRESS_2);
-  const operatorTo = newAddress('operatorTo', MOCK_ADDRESS_1);
-  idx = newI32('idx', 1);
-
-  event.parameters.push(operatorFrom);
-  event.parameters.push(operatorTo);
-  event.parameters.push(idx);
-
-  assert.fieldEquals('RedeemOperation', MOCK_ADDRESS_2, 'idx', '2');
-  assert.fieldEquals('RedeemOperation', MOCK_ADDRESS_1, 'idx', '1');
-
-  handleRearrangeRedeem(event);
-
-  assert.fieldEquals('RedeemOperation', MOCK_ADDRESS_2, 'idx', '1');
-  assert.fieldEquals('RedeemOperation', MOCK_ADDRESS_1, 'idx', '1');
-});
-
-test('handleFinishRedeem updates the RedeemOperation', () => {
-  clearStore();
-
-  const depositOperation = new RedeemOperation(MOCK_ADDRESS_1);
-  depositOperation.save();
-
-  let mockEvent = newMockEvent();
-  const event = new FinishRedeemStable(
-    mockEvent.address,
-    mockEvent.logIndex,
-    mockEvent.transactionLogIndex,
-    mockEvent.logType,
-    mockEvent.block,
-    mockEvent.transaction,
-    mockEvent.parameters,
-  );
-  event.parameters = new Array();
-
-  const operator = newAddress('operator', MOCK_ADDRESS_1);
-
-  event.parameters.push(operator);
-
-  handleFinishRedeem(event);
-
-  assert.notInStore('RedeemOperation', MOCK_ADDRESS_1);
-});
-
-test('handleTreasuryUpdated updates the treasury', () => {
-  clearStore();
-
-  let mockEvent = newMockEvent();
-  const event = new TreasuryUpdated(
-    mockEvent.address,
-    mockEvent.logIndex,
-    mockEvent.transactionLogIndex,
-    mockEvent.logType,
-    mockEvent.block,
-    mockEvent.transaction,
-    mockEvent.parameters,
-  );
-  event.parameters = new Array();
-
-  const treasury = newAddress('treasury', MOCK_ADDRESS_1);
-  event.parameters.push(treasury);
-
-  // create vault
-  const vault = new Vault(mockEvent.address.toHexString());
-  vault.save();
-
-  handleTreasuryUpdated(event);
-
-  assert.fieldEquals(
-    'Vault',
-    mockEvent.address.toHexString(),
-    'treasury',
-    MOCK_ADDRESS_1,
-  );
-});
-
-test('handleSponsored creates a Sponsor', () => {
-  clearStore();
-
-  let mockEvent = newMockEvent();
-  const event = new Sponsored(
-    mockEvent.address,
-    mockEvent.logIndex,
-    mockEvent.transactionLogIndex,
-    mockEvent.logType,
-    mockEvent.block,
-    mockEvent.transaction,
-    mockEvent.parameters,
-  );
-  event.parameters = new Array();
-
-  const idParam = newI32('id', 1);
-  const amount = newI32('amount', 1);
-  const depositor = newAddress('depositor', MOCK_ADDRESS_1);
-  const lockedUntil = newI32('lockedUntil', 1);
-  const burned = newBool('burned', false);
-
-  event.parameters.push(idParam);
-  event.parameters.push(amount);
-  event.parameters.push(depositor);
-  event.parameters.push(lockedUntil);
-  event.parameters.push(burned);
-
-  handleSponsored(event);
-
-  assert.fieldEquals('Sponsor', '1', 'amount', '1');
-  assert.fieldEquals('Sponsor', '1', 'depositor', MOCK_ADDRESS_1);
-  assert.fieldEquals('Sponsor', '1', 'burned', 'false');
-});
-
-test('handleUnsponsored removes a Sponsor by marking as burned', () => {
-  clearStore();
-
-  const sponsor = new Sponsor('1');
-  sponsor.burned = false;
-  sponsor.save();
-
-  let mockEvent = newMockEvent();
-  const event = new Unsponsored(
-    mockEvent.address,
-    mockEvent.logIndex,
-    mockEvent.transactionLogIndex,
-    mockEvent.logType,
-    mockEvent.block,
-    mockEvent.transaction,
-    mockEvent.parameters,
-  );
-  event.parameters = new Array();
-
-  const idParam = newI32('id', 1);
-
-  event.parameters.push(idParam);
-
-  handleUnsponsored(event);
-
-  assert.fieldEquals('Sponsor', '1', 'burned', 'true');
-});
-
-test('handleDepositMinted creates a Deposit', () => {
+test('handleDepositMinted updates the Foundation', () => {
   clearStore();
 
   let mockEvent = newMockEvent();
@@ -436,76 +79,14 @@ test('handleDepositMinted creates a Deposit', () => {
   const vault = new Vault(mockEvent.address.toHexString());
   vault.save();
 
-  const idParam = newI32('id', 1);
+  let idParam = newI32('id', 1);
   const groupId = newI32('groupId', 1);
-  const amount = newI32('amount', 1);
-  const shares = newI32('shares', 1);
-  const depositor = newAddress('depositor', MOCK_ADDRESS_1);
-  const claimer = newAddress('claimer', MOCK_ADDRESS_1);
-  const lockedUntil = newI32('lockedUntil', 1);
-  const data = newBytes('data', Bytes.empty());
-  const name = newString('name', 'Foundation');
-
-  event.parameters.push(idParam);
-  event.parameters.push(groupId);
-  event.parameters.push(amount);
-  event.parameters.push(shares);
-  event.parameters.push(depositor);
-  event.parameters.push(claimer);
-  event.parameters.push(claimer);
-  event.parameters.push(lockedUntil);
-  event.parameters.push(data);
-  event.parameters.push(name);
-
-  handleDepositMinted(event);
-
-  assert.fieldEquals('Deposit', '1', 'amount', '1');
-  assert.fieldEquals('Deposit', '1', 'depositor', MOCK_ADDRESS_1);
-  assert.fieldEquals('Deposit', '1', 'claimer', MOCK_ADDRESS_1);
-  assert.fieldEquals('Claimer', MOCK_ADDRESS_1, 'principal', '1');
-  assert.fieldEquals('Claimer', MOCK_ADDRESS_1, 'depositsIds', '[1]');
-
-  const foundationId = `${vault.id}-1`;
-  assert.fieldEquals('Foundation', foundationId, 'name', 'Foundation');
-  assert.fieldEquals('Foundation', foundationId, 'owner', MOCK_ADDRESS_1);
-  assert.fieldEquals('Foundation', foundationId, 'vault', vault.id);
-  assert.fieldEquals('Foundation', foundationId, 'amountDeposited', '1');
-  assert.fieldEquals('Foundation', foundationId, 'lockedUntil', '1');
-  assert.fieldEquals(
-    'Foundation',
-    foundationId,
-    'createdAt',
-    event.block.timestamp.toString(),
-  );
-});
-
-test("handleDepositMinted uses the last event's name as the Foundation's name", () => {
-  clearStore();
-
-  let mockEvent = newMockEvent();
-  let event = new DepositMinted(
-    mockEvent.address,
-    mockEvent.logIndex,
-    mockEvent.transactionLogIndex,
-    mockEvent.logType,
-    mockEvent.block,
-    mockEvent.transaction,
-    mockEvent.parameters,
-  );
-  event.parameters = new Array();
-
-  const vault = new Vault(mockEvent.address.toHexString());
-  vault.save();
-
-  const idParam = newI32('id', 1);
-  const groupId = newI32('groupId', 1);
-  const amount = newI32('amount', 1);
-  const shares = newI32('shares', 1);
-  const depositor = newAddress('depositor', MOCK_ADDRESS_1);
-  const claimer = newAddress('claimer', MOCK_ADDRESS_1);
-  const claimerId = newAddress('claimerId', MOCK_ADDRESS_1);
-  const lockedUntil = newI32('lockedUntil', 1);
-  const data = newBytes('data', Bytes.empty());
+  let amount = newI32('amount', 10);
+  let shares = newI32('shares', 10);
+  let depositor = newAddress('depositor', MOCK_ADDRESS_1);
+  let claimer = newAddress('claimer', MOCK_ADDRESS_1);
+  let lockedUntil = newI32('lockedUntil', 1);
+  let data = newBytes('data', Bytes.empty());
   let name = newString('name', 'Foundation');
 
   event.parameters.push(idParam);
@@ -514,78 +95,32 @@ test("handleDepositMinted uses the last event's name as the Foundation's name", 
   event.parameters.push(shares);
   event.parameters.push(depositor);
   event.parameters.push(claimer);
-  event.parameters.push(claimerId);
-  event.parameters.push(lockedUntil);
-  event.parameters.push(data);
-  event.parameters.push(name);
-
-  handleDepositMinted(event);
-
-  const foundationId = `${mockEvent.address.toHexString()}-1`;
-  assert.fieldEquals('Foundation', foundationId, 'name', 'Foundation');
-
-  // Sending another DepositMinted that updates the name
-
-  mockEvent = newMockEvent();
-  event = new DepositMinted(
-    mockEvent.address,
-    mockEvent.logIndex,
-    mockEvent.transactionLogIndex,
-    mockEvent.logType,
-    mockEvent.block,
-    mockEvent.transaction,
-    mockEvent.parameters,
-  );
-  event.parameters = new Array();
-
-  name = newString('name', 'Updated Foundation Name');
-
-  event.parameters.push(idParam);
-  event.parameters.push(groupId);
-  event.parameters.push(amount);
-  event.parameters.push(shares);
-  event.parameters.push(depositor);
   event.parameters.push(claimer);
-  event.parameters.push(claimerId);
   event.parameters.push(lockedUntil);
   event.parameters.push(data);
   event.parameters.push(name);
 
   handleDepositMinted(event);
 
+  const foundationId = `${vault.id}-1`;
+
+  assert.fieldEquals('Foundation', foundationId, 'name', 'Foundation');
+  assert.fieldEquals('Foundation', foundationId, 'owner', MOCK_ADDRESS_1);
+  assert.fieldEquals('Foundation', foundationId, 'vault', vault.id);
+  assert.fieldEquals('Foundation', foundationId, 'amountDeposited', '10');
+  assert.fieldEquals('Foundation', foundationId, 'shares', '10');
+  assert.fieldEquals('Foundation', foundationId, 'lockedUntil', '1');
   assert.fieldEquals(
     'Foundation',
     foundationId,
-    'name',
-    'Updated Foundation Name',
+    'createdAt',
+    event.block.timestamp.toString(),
   );
-});
 
-test("handleDepositWithdrawn doesn't remove a Deposit for partial withdraws", () => {
-  clearStore();
+  // Second deposit
 
-  let mockEvent = newMockEvent();
-
-  const claimer = new Claimer('1');
-  claimer.save();
-
-  const deposit = new Deposit('1');
-  deposit.burned = false;
-  deposit.amount = BigInt.fromI32(10);
-  deposit.lockedUntil = BigInt.fromI32(1);
-  deposit.shares = BigInt.fromI32(10);
-  deposit.claimer = '1';
-  deposit.foundation = '1';
-  deposit.save();
-
-  const vault = new Vault(mockEvent.address.toHexString());
-  vault.save();
-
-  const foundation = new Foundation('1');
-  foundation.vault = mockEvent.address.toHexString();
-  foundation.save();
-
-  const event = new DepositWithdrawn(
+  mockEvent = newMockEvent();
+  const event2 = new DepositMinted(
     mockEvent.address,
     mockEvent.logIndex,
     mockEvent.transactionLogIndex,
@@ -594,227 +129,42 @@ test("handleDepositWithdrawn doesn't remove a Deposit for partial withdraws", ()
     mockEvent.transaction,
     mockEvent.parameters,
   );
-  event.parameters = new Array();
+  event2.parameters = new Array();
 
-  event.parameters.push(newI32('id', 1));
-  event.parameters.push(newI32('shares', 5));
-  event.parameters.push(newI32('amount', 5));
-  event.parameters.push(newAddress('to', MOCK_ADDRESS_1));
-  event.parameters.push(newBool('burned', false));
+  idParam = newI32('id', 1);
+  amount = newI32('amount', 20);
+  shares = newI32('shares', 15);
+  depositor = newAddress('depositor', MOCK_ADDRESS_1);
+  claimer = newAddress('claimer', MOCK_ADDRESS_1);
+  lockedUntil = newI32('lockedUntil', 2);
+  data = newBytes('data', Bytes.empty());
+  name = newString('name', 'Foundation 2');
 
-  handleDepositWithdrawn(event);
+  event2.parameters.push(idParam);
+  event2.parameters.push(groupId);
+  event2.parameters.push(amount);
+  event2.parameters.push(shares);
+  event2.parameters.push(depositor);
+  event2.parameters.push(claimer);
+  event2.parameters.push(claimer);
+  event2.parameters.push(lockedUntil);
+  event2.parameters.push(data);
+  event2.parameters.push(name);
 
-  assert.fieldEquals('Deposit', '1', 'burned', 'false');
-  assert.fieldEquals('Deposit', '1', 'shares', '5');
-  assert.fieldEquals('Deposit', '1', 'amount', '5');
-  assert.fieldEquals('Foundation', '1', 'amountDeposited', '0');
-});
+  handleDepositMinted(event2);
 
-test('handleDepositWithdrawn removes a Deposit by marking as burned', () => {
-  clearStore();
-
-  let mockEvent = newMockEvent();
-
-  const claimer = new Claimer('1');
-  claimer.save();
-
-  const deposit = new Deposit('1');
-  deposit.burned = false;
-  deposit.amount = BigInt.fromI32(1);
-  deposit.lockedUntil = BigInt.fromI32(1);
-  deposit.shares = BigInt.fromI32(1);
-  deposit.claimer = '1';
-  deposit.foundation = '1';
-  deposit.save();
-
-  const vault = new Vault(mockEvent.address.toHexString());
-  vault.save();
-
-  const foundation = new Foundation('1');
-  foundation.vault = mockEvent.address.toHexString();
-  foundation.save();
-
-  const event = new DepositWithdrawn(
-    mockEvent.address,
-    mockEvent.logIndex,
-    mockEvent.transactionLogIndex,
-    mockEvent.logType,
-    mockEvent.block,
-    mockEvent.transaction,
-    mockEvent.parameters,
+  assert.fieldEquals('Foundation', foundationId, 'name', 'Foundation 2');
+  assert.fieldEquals('Foundation', foundationId, 'owner', MOCK_ADDRESS_1);
+  assert.fieldEquals('Foundation', foundationId, 'vault', vault.id);
+  assert.fieldEquals('Foundation', foundationId, 'amountDeposited', '30');
+  assert.fieldEquals('Foundation', foundationId, 'shares', '25');
+  assert.fieldEquals('Foundation', foundationId, 'lockedUntil', '2');
+  assert.fieldEquals(
+    'Foundation',
+    foundationId,
+    'createdAt',
+    event.block.timestamp.toString(),
   );
-  event.parameters = new Array();
-
-  event.parameters.push(newI32('id', 1));
-  event.parameters.push(newI32('shares', 1));
-  event.parameters.push(newI32('amount', 1));
-  event.parameters.push(newAddress('to', MOCK_ADDRESS_1));
-  event.parameters.push(newBool('burned', true));
-
-  handleDepositWithdrawn(event);
-
-  assert.fieldEquals('Deposit', '1', 'burned', 'true');
-  assert.fieldEquals('Deposit', '1', 'shares', '0');
-  assert.fieldEquals('Deposit', '1', 'amount', '0');
-  assert.fieldEquals('Foundation', '1', 'amountDeposited', '0');
-});
-
-test('handleYieldClaimed reduces shares from Deposits and creates Donations', () => {
-  clearStore();
-
-  let mockEvent = newMockEvent();
-
-  // Create deposits
-  createDeposit('1', 50, false, '1', '1', 1, 50);
-  createDeposit('2', 100, false, '1', '1', 1, 100);
-
-  // Create vault
-  const vault = new Vault(mockEvent.address.toHexString());
-  vault.treasury = Address.fromString(TREASURY_ADDRESS);
-  vault.save();
-
-  // Create claimer
-  const claimer = new Claimer(MOCK_ADDRESS_1);
-  claimer.vault = mockEvent.address.toHexString();
-  claimer.depositsIds = ['1', '2'];
-  claimer.save();
-
-  // Create foundation
-  const foundation = new Foundation('1');
-  foundation.vault = mockEvent.address.toHexString();
-  foundation.save();
-
-  const event = new YieldClaimed(
-    mockEvent.address,
-    mockEvent.logIndex,
-    mockEvent.transactionLogIndex,
-    mockEvent.logType,
-    mockEvent.block,
-    mockEvent.transaction,
-    mockEvent.parameters,
-  );
-  event.parameters = new Array();
-
-  event.parameters.push(newAddress('claimerId', MOCK_ADDRESS_1));
-  event.parameters.push(newAddress('to', TREASURY_ADDRESS));
-  event.parameters.push(newI32('amount', 150));
-  event.parameters.push(newI32('burnedShares', 75));
-  event.parameters.push(newI32('perfFee', 0));
-
-  handleYieldClaimed(event);
-
-  assert.fieldEquals('Deposit', '1', 'shares', '25');
-  assert.fieldEquals('Deposit', '2', 'shares', '50');
-
-  assert.fieldEquals('Donation', donationId(mockEvent, '0'), 'amount', '50');
-  assert.fieldEquals('Donation', donationId(mockEvent, '1'), 'amount', '100');
-
-  clearStore();
-});
-
-test('handleYieldClaimed takes the performance fee into account', () => {
-  clearStore();
-
-  let mockEvent = newMockEvent();
-
-  // Create deposits
-  createDeposit('1', 50, false, '1', '1', 1, 50);
-  createDeposit('2', 100, false, '1', '1', 1, 100);
-
-  // Create vault
-  const vault = new Vault(mockEvent.address.toHexString());
-  vault.treasury = Address.fromString(TREASURY_ADDRESS);
-  vault.save();
-
-  // Create claimer
-  const claimer = new Claimer(MOCK_ADDRESS_1);
-  claimer.vault = mockEvent.address.toHexString();
-  claimer.depositsIds = ['1', '2'];
-  claimer.save();
-
-  // Create foundation
-  const foundation = new Foundation('1');
-  foundation.vault = mockEvent.address.toHexString();
-  foundation.save();
-
-  const event = new YieldClaimed(
-    mockEvent.address,
-    mockEvent.logIndex,
-    mockEvent.transactionLogIndex,
-    mockEvent.logType,
-    mockEvent.block,
-    mockEvent.transaction,
-    mockEvent.parameters,
-  );
-  event.parameters = new Array();
-
-  event.parameters.push(newAddress('claimerId', MOCK_ADDRESS_1));
-  event.parameters.push(newAddress('to', TREASURY_ADDRESS));
-  event.parameters.push(newI32('amount', 120));
-  event.parameters.push(newI32('burnedShares', 75));
-  event.parameters.push(newI32('perfFee', 30));
-
-  handleYieldClaimed(event);
-
-  assert.fieldEquals('Claimer', MOCK_ADDRESS_1, 'claimed', '120');
-
-  assert.fieldEquals('Deposit', '1', 'shares', '25');
-  assert.fieldEquals('Deposit', '2', 'shares', '50');
-
-  assert.fieldEquals('Donation', donationId(mockEvent, '0'), 'amount', '40');
-  assert.fieldEquals('Donation', donationId(mockEvent, '1'), 'amount', '80');
-
-  clearStore();
-});
-
-test("handleYieldClaimed doesn't create donations if the deposits are not to the treasury", () => {
-  clearStore();
-
-  let mockEvent = newMockEvent();
-
-  // Create deposits
-  createDeposit('1', 50, false, '1', '1', 1, 50);
-  createDeposit('2', 100, false, '1', '1', 1, 100);
-
-  // Create vault
-  const vault = new Vault(mockEvent.address.toHexString());
-  vault.treasury = Address.fromString(TREASURY_ADDRESS);
-  vault.save();
-
-  // Create claimer
-  const claimer = new Claimer(MOCK_ADDRESS_2);
-  claimer.vault = mockEvent.address.toHexString();
-  claimer.depositsIds = ['1', '2'];
-  claimer.save();
-
-  // Create foundation
-  const foundation = new Foundation('1');
-  foundation.vault = mockEvent.address.toHexString();
-  foundation.save();
-
-  const event = new YieldClaimed(
-    mockEvent.address,
-    mockEvent.logIndex,
-    mockEvent.transactionLogIndex,
-    mockEvent.logType,
-    mockEvent.block,
-    mockEvent.transaction,
-    mockEvent.parameters,
-  );
-  event.parameters = new Array();
-
-  event.parameters.push(newAddress('claimerId', MOCK_ADDRESS_2));
-  event.parameters.push(newAddress('to', MOCK_ADDRESS_1));
-  event.parameters.push(newI32('amount', 150));
-  event.parameters.push(newI32('burnedShares', 75));
-  event.parameters.push(newI32('perfFee', 0));
-
-  handleYieldClaimed(event);
-
-  assert.notInStore('Donation', donationId(mockEvent, '0'));
-  assert.notInStore('Donation', donationId(mockEvent, '1'));
-
-  clearStore();
 });
 
 test('handleYieldClaimed handles scenarios where only one of the deposits generated yield', () => {
@@ -910,51 +260,3 @@ test('handleYieldClaimed handles scenarios where the yield is not proportional t
   assert.fieldEquals('Donation', donationId(mockEvent, '0'), 'amount', '99');
   assert.fieldEquals('Donation', donationId(mockEvent, '1'), 'amount', '48');
 });
-
-function newBytes(name: string, value: Bytes): ethereum.EventParam {
-  return new ethereum.EventParam(name, ethereum.Value.fromBytes(value));
-}
-
-function newI32(name: string, value: i32): ethereum.EventParam {
-  return new ethereum.EventParam(name, ethereum.Value.fromI32(value));
-}
-
-function newBool(name: string, value: bool): ethereum.EventParam {
-  return new ethereum.EventParam(name, ethereum.Value.fromBoolean(value));
-}
-
-function newAddress(name: string, value: string): ethereum.EventParam {
-  return new ethereum.EventParam(
-    name,
-    ethereum.Value.fromAddress(Address.fromString(value)),
-  );
-}
-
-function newString(name: string, value: string): ethereum.EventParam {
-  return new ethereum.EventParam(name, ethereum.Value.fromString(value));
-}
-
-function donationId(event: ethereum.Event, id: string): string {
-  return (
-    event.transaction.hash.toHex() + '-' + event.logIndex.toString() + '-' + id
-  );
-}
-
-function createDeposit(
-  id: string,
-  amount: i32,
-  burned: bool,
-  claimer: string,
-  foundation: string,
-  lockedUntil: i32,
-  shares: i32,
-): void {
-  const deposit = new Deposit(id);
-  deposit.amount = BigInt.fromI32(amount);
-  deposit.burned = burned;
-  deposit.claimer = claimer;
-  deposit.foundation = foundation;
-  deposit.lockedUntil = BigInt.fromI32(lockedUntil);
-  deposit.shares = BigInt.fromI32(shares);
-  deposit.save();
-}
